fix storing 64-bit value from vstore
We can't just always reset sv->c.i in load_symofs since if we are called
from the USING_TWO_WORDS branch in vstore. The sv->c.i is required to
generate the second store. If cleared the second store will be located
to a wrong position (will be fp+4, which is not suitable for a stack
value). This value got cleared in privious version to allow emit_Sx work
in store, but we can make use of it by only load the upper part of
sv->c.i in load_symofs